Math: Use table lookup method to calucate sin.
And we use integer phase instead of float num, which greatly improves performance.

+++ include/player_math.h
@@ -1,5 +1,5 @@
 #define PI 3.141592653589793
 #define PI_DIV_2 PI/2
 #define PI_CROSS_2 PI*2
-float fastsin(float x);
+float fastsin(int x);
 void sintab_init(void);
